requirements.txt
Flask
Flask_SQLAlchemy
Flask_Script
cymysql
flask-sqlacodegen